chore: makes sure data is saved without rounding errors and more

+++ src/main/java/org/acme/domain/dtos/PresentInstallmentDto.java
@@ -1,12 +1,14 @@
 package org.acme.domain.dtos;
 
+import java.math.BigDecimal;
+
 import org.acme.domain.models.Installment;
 
 public record PresentInstallmentDto(
     int number,
-    Double amortization,
-    Double interest,
-    Double value
+    BigDecimal amortization,
+    BigDecimal interest,
+    BigDecimal value
 ) {
     public static PresentInstallmentDto from(Installment installment) {
         return new PresentInstallmentDto(
